refactor: equals를 Player 클래스에서만 사용하도록 이동
