+++ src/main/java/domain/participant/Participant.java
@@ -3,7 +3,6 @@ package domain.participant;
 import domain.card.Card;
 import domain.card.CardBundle;
 import java.util.List;
-import java.util.Objects;
 
 public abstract class Participant {
 
@@ -39,17 +38,4 @@ public abstract class Participant {
         return cardBundle.toDisplay();
     }
 
-    @Override
-    public boolean equals(Object o) {
-        if (o == null || getClass() != o.getClass()) {
-            return false;
-        }
-        Participant that = (Participant) o;
-        return Objects.equals(name, that.name) && Objects.equals(cardBundle, that.cardBundle);
-    }
-
-    @Override
-    public int hashCode() {
-        return Objects.hash(name, cardBundle);
-    }
 }
