+++ src/main/java/domain/participant/Player.java
@@ -1,5 +1,7 @@
 package domain.participant;
 
+import java.util.Objects;
+
 public class Player extends Participant{
 
     private Player(ParticipantName name) {
@@ -8,5 +10,19 @@ public class Player extends Participant{
 
     public static Player from(ParticipantName name) {
         return new Player(name);
+    }
+
+    @Override
+    public boolean equals(Object o) {
+        if (o == null || getClass() != o.getClass()) {
+            return false;
+        }
+        Participant that = (Participant) o;
+        return Objects.equals(name, that.name) && Objects.equals(cardBundle, that.cardBundle);
+    }
+
+    @Override
+    public int hashCode() {
+        return Objects.hash(name, cardBundle);
     }
 }
